refactoring: moved low-level file operations to a dedicated module
